+++ frontend/src/components/pages/search/Search.tsx
@@ -53,10 +53,7 @@ export const SearchUI: React.FC<Props> = ({ language }) => {
   const { filtersState, setFilterSelectedOptions, resetFilters } = useFilter();
 
   const { subMenuState, selectFilter, hideSubMenu, currentFilterId } = useFilterSubMenu();
-  const { menuState, displayMenu, hideMenu, filtersList } = useFilterMenu(
-    filtersState,
-    selectFilter,
-  );
+  const { menuState, displayMenu, hideMenu, filtersList } = useFilterMenu(selectFilter);
 
   const { bboxState, handleMoveMap } = useBbox();