+++ frontend/src/modules/details/adapter.ts
@@ -13,6 +13,7 @@ import { dataUnits } from 'modules/results/adapter';
 import { TrekResult } from 'modules/results/interface';
 import { formatDistance } from 'modules/results/utils';
 import { SensitiveArea } from 'modules/sensitiveArea/interface';
+import { SignageDictionary } from 'modules/signage/interface';
 import { SourceDictionnary } from 'modules/source/interface';
 import { TouristicContent } from 'modules/touristicContent/interface';
 import { getAttachments } from 'modules/utils/adapter';
@@ -40,6 +41,7 @@ export const adaptResults = ({
   children,
   childrenGeometry,
   sensitiveAreas,
+  signage,
   reservation,
   trekRating,
   trekRatingScale,
@@ -61,6 +63,7 @@ export const adaptResults = ({
   children: TrekResult[];
   childrenGeometry: TrekChildGeometry[];
   sensitiveAreas: SensitiveArea[];
+  signage: SignageDictionary | null;
   reservation: Reservation | null;
   trekRating: TrekRatingChoices;
   trekRatingScale: TrekRatingScale[];
@@ -155,7 +158,7 @@ export const adaptResults = ({
         geometry: childrenGeometry.find(childGeometry => childGeometry.id === `${child.id}`),
       })),
       sensitiveAreas,
-      webLinks: rawDetailsProperties.web_links,
+      webLinks: rawDetailsProperties.web_links ?? null,
       elevationAreaUrl: rawDetailsProperties.elevation_area_url,
       altimetricProfileUrl: rawDetailsProperties.altimetric_profile,
       length2d: rawDetailsProperties.length_2d,
@@ -169,6 +172,7 @@ export const adaptResults = ({
           };
         }) ?? [],
       ratingsDescription: rawDetailsProperties.ratings_description ?? '',
+      signage,
     };
   } catch (e) {
     console.error('Error in details/adapter', e);