+++ frontend/src/modules/details/interface.ts
@@ -20,6 +20,7 @@ import { InformationDesk } from 'modules/informationDesk/interface';
 import { Label } from 'modules/label/interface';
 import { TrekResult } from 'modules/results/interface';
 import { SensitiveArea } from 'modules/sensitiveArea/interface';
+import { SignageDictionary } from 'modules/signage/interface';
 import { TrekRatingWithScale } from '../trekRating/interface';
 
 export interface RawDetails {
@@ -182,6 +183,7 @@ export interface Details extends DetailsHtml {
   length2d: number;
   reservation: Reservation | null;
   reservation_id: string | null;
+  signage: SignageDictionary | null;
 }
 
 export interface WebLink {