+++ docs/changelog.md
@@ -1,13 +1,41 @@
 # Changelog
 
+3.8.4 (2022-05-31)
+------------------
+
+**🚀 New features**
+
+* Allow to override ``manifest.json`` properties (#661)
+
+**🐛 Fixes**
+
+* Fix card display for iOS mobile (#645)
+* Fix truncated icon name if it is too long (#658)
+* Display Sensitive Areas when their geometry is MultiPolygon (#655)
+* Fix trek detail page if there are no signages endpoint in API (#660)
+* Fix trek and outdoor site detail page if there are no weblinks (#660)
+
+3.8.3 (2022-05-24)
+------------------
+
+**🚀 New features**
+
+* Add published signages on treks and outdoor maps (#408)
+* Add ``groupTreksAndOutdoorFilters`` setting to be able to group Treks and Outdoor filters on search page (#656)
+
+**🐛 Fixes**
+
+* Fix flickering of Outdoor, services and events maps
+* Refactoring and improvements of anchor system in detail pages
+
 3.8.2 (2022-04-25)
 ------------------
 
 **🐛 Fixes**
 
-- Fix HTML interpretation of new "Accessibility" fields (#536 / #638)
-- Don't display label filter (Others) in trek filters if no label defined as filter (#418)
-- Improve outdoor subobjects display on site maps with adding it in layer control (#542)
+* Fix HTML interpretation of new "Accessibility" fields (#536 / #638)
+* Don't display label filter (Others) in trek filters if no label defined as filter (#418)
+* Improve outdoor subobjects display on site maps with adding it in layer control (#542)
 
 3.8.1 (2022-04-25)
 ------------------